requirements.txt
# Google Gemini API
google-generativeai>=0.8.0

# Structured output enforcement
instructor>=1.4.0
pydantic>=2.5.0

# PDF parsing libraries
pypdf>=4.0.0
pdfplumber>=0.10.0
pymupdf>=1.24.0

# Data processing and Excel output
pandas>=2.1.0
openpyxl>=3.1.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Progress bars
tqdm>=4.66.0

# HTTP requests with retry logic
tenacity>=8.2.0

# Jupyter notebook
jupyter>=1.0.0
ipykernel>=6.27.0

# Logging and utilities
python-dotenv>=1.0.0
